requirements.txt
requests
allure-behave
behave
pyyaml
assertpy
paramiko
pydash
selenium
webdriver-manager
